test: remove duplicate ElementJacksonTest, fold unique cases into ElementTest
ElementJacksonTest was a near-identical fork of ElementTest added during
the Jackson migration. Folded the parts that were not already covered into
ElementTest: enable setStyle() (was missing @Test annotation), assert
parsed height in setMultipleStylesAsAttribute, extend
setComplexStylesAsAttribute with calc/hsl/url/cross-fade/image/var cases
and have testStyleAttribute return Style, and add callFunctionWithBean.
